requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
Pillow==11.1.0
python-decouple==3.8
twilio==9.4.0
stripe==11.5.0
qrcode==8.0
reportlab==4.2.5
django-filter==24.3
